Refactored the UI layout. Created independent login and register form components
