+++ src/App.tsx
@@ -1,21 +1,16 @@
 /** @jsxImportSource @emotion/react */
 import { useState } from 'react';
 import { Logo } from './components/logo';
-import UserPasswordForm from './components/user-password-form';
-import { Button, CircleButton, Dialog } from './components/lib/UI';
-import { DialogContent, DialogTitle } from '@mui/material';
+import { Button } from './components/lib/UI';
+import LoginDialog from './components/layout/login-dialog';
+import RegisterDialog from './components/layout/register-dialog';
 
 type Modals = 'none' | 'login' | 'register';
-
-type MuiDialogCloseReason = 'backdropClick' | 'escapeKeyDown';
 
 const App: React.FC = () => {
   const [openModal, setOpenModal] = useState<Modals>('none');
 
-  const handleDialogClose = (
-    event: MouseEvent,
-    reason: MuiDialogCloseReason
-  ) => {
+  const handleDialogClose = () => {
     setOpenModal('none');
   };
 
@@ -44,46 +39,21 @@ const App: React.FC = () => {
           Register
         </Button>
       </div>
-      <Dialog
-        open={openModal === 'login'}
+
+      <LoginDialog
+        isOpen={openModal === 'login'}
+        onSubmit={(username, password) =>
+          console.log('login: ', username, password)
+        }
         onClose={handleDialogClose}
-        PaperProps={{ style: { padding: '1rem 1rem 3rem' } }}
-      >
-        <CircleButton
-          onClick={() => setOpenModal('none')}
-          css={{ marginLeft: 'auto' }}
-        >
-          &times;
-        </CircleButton>
-        <DialogTitle>Login</DialogTitle>
-        <DialogContent>
-          <UserPasswordForm
-            onSubmit={(username, password) =>
-              console.log('login: ', username, password)
-            }
-          />
-        </DialogContent>
-      </Dialog>
-      <Dialog
-        open={openModal === 'register'}
+      />
+      <RegisterDialog
+        isOpen={openModal === 'register'}
+        onSubmit={(username, password) =>
+          console.log('register: ', username, password)
+        }
         onClose={handleDialogClose}
-        PaperProps={{ style: { padding: '1rem 1rem 3rem' } }}
-      >
-        <CircleButton
-          onClick={() => setOpenModal('none')}
-          css={{ marginLeft: 'auto' }}
-        >
-          &times;
-        </CircleButton>
-        <DialogTitle>Register</DialogTitle>
-        <DialogContent>
-          <UserPasswordForm
-            onSubmit={(username, password) =>
-              console.log('register: ', username, password)
-            }
-          />
-        </DialogContent>
-      </Dialog>
+      />
     </div>
   );
 };
